[VertexAI] Add initial support to export covered API (#6749)
The server produces a discovery document with the details of the API
surface.

https://aiplatform.googleapis.com/$discovery/rest?version=v1beta1

This change introduces code that can generate similar a description of
the API covered by the SDK. This will enable us to track difference
between both.

In a follow up PR we can implement the logic to fully export the
surface.

---------

Co-authored-by: Daymon <17409137+daymxn@users.noreply.github.com>