Store registered context for sync task unregistrations (#6752)
For issue #6558, this is an attempt at fixing the
IllegalArgumentException by ensuring that the context we use for
registering the SyncTask is the same context we use to unregister the
task. Race conditions dont seem like a culprit here since unregister is
only triggered by the Receiver itself, which should be only executed
synchronously on the main thread.