Add transform action to remove package prefix in toc files (#6787)
Per [b/379093944](https://b.corp.google.com/issues/379093944),

This adds a transform action to remove the `com.google.` package prefix
in the TOC files, as it pollutes the visual landscape for consumers.
Long-term, this action will be replaced by a parameter offered via newer
versions of dackka- but we don't currently have the bandwidth for such
an upgrade. A tracking bug has been left with the implementation to
migrate to said parameter whenever we have the time to upgrade.