[VertexAI] Log warning for unsupported model names (#6805)
Added a warning message to the initializers of GenerativeModel and
ImagenModel that is logged when the provided model name does not start
with the expected prefix ("gemini-" for GenerativeModel and "imagen-"
for ImagenModel). The warning message includes a link to the
documentation for supported models.

Note: No error is thrown in case the naming scheme is changed in the
future, though we would want to update the logic/message at that time.

Related iOS PR https://github.com/firebase/firebase-ios-sdk/pull/14610