Add basic Vertex Java compilation tests (#6810)
This is a starting point for compilation testing, broadly using most
symbols Vertex exposes to Java users, lightly validating usability and
structure. As a note, there are some builder patterns in Vertex that
don't work as expected from Java and were omitted, as fixing that would
be a breaking change.