Improve the integration test coverage for online vs offline comparisons. (#6841)
This PR improves the way `checkOnlineAndOfflineResultsMatch` test util
function was written.

It does so by first pre-populating the cache with the entire collection,
then performing the query from cache (this results in a full collection
scan), then performing the query from the server, and then performing
the query from cache again (this results in using
`performQueryUsingRemoteKeys`. It then ensures that all of these 3
results are the same and equal to the expected results.

#no-changelog