Enable multimodal response generation in android (#6901)
This change enables use of multiple modalities when calling
generateContent from the model.
This change adds a new field into the GenerationConfig to specify the
responseModalities which would be sent to the server. Also for easier
response handling it exposes a function inlineDataParts which should
return all the data sent back by the model.

---------

Co-authored-by: Daymon <17409137+daymxn@users.noreply.github.com>