Migrate LiveContentResponse.Status to properties (#6906)
Per [b/410855548](https://b.corp.google.com/issues/410855548),

This removes the `LiveContentResponse.Status` class, and instead nests
the status as corresponding fields directly on `LiveContentResponse`.
This is done to ensure we can support the model returning multiple
statuses, since the model does not define them as exclusive.

Note that proto alignment efforts will be in a separate PR. This PR is
_only_ for migrating `Status` to properties.