Exclude protolite well known types from version bump. (#6909)
Per [b/413403910](https://b.corp.google.com/issues/413403910),

This adds filtering logic to not bump the version of
`protolite-well-known-types` when it releases. This solves the issue of
the `version` and `latestReleasedVersion` not aligning (which must be
enforced due to
[b/285892320](https://buganizer.corp.google.com/issues/285892320)).

More specifically, this PR adds a toggle for the default `bump` behavior
in `VersionBumpTask`. We then set this to `false` when the project's
`artifactId` is `protolite-well-known-types`.

This allows us to still [automatically] update the
`latestReleasedVersion` (should we ever release well known types again),
without bumping the `version`.

The `VersionBumpTask` docs have been updated to accommodate this.