Fix AI test timeout (#6917)
Per [b/414406390](https://b.corp.google.com/issues/414406390),

This fixes the issue with long text golden files timing out the tests.
This occurred because the text contents were too large for the channel's
buffer- so the channel entered a suspension point until the content was
read. Unfortunately, this suspension point blocked the test thread. To
solve this, we do the same thing we do for `goldenStreamingFile`- we
write to the channel in a separate coroutine scope. This allows Kotlin
to properly recognize that the suspension point within the writer should
allow for the test thread to continue.