Align LiveSeverMessage related protos (#6910)
Per [b/410856389](https://b.corp.google.com/issues/410856389),

This removes the `LiveContentResponse` and provides a
`LiveServerMessage` in its place that aligns with the proto, and the
other SDKs. Documentation is included for everything added.

Additionally, this PR fixes the following:

- [b/410040441](https://b.corp.google.com/issues/410040441) -> Add
support for id field on functions
- [b/413437559](https://b.corp.google.com/issues/413437559) -> Add
support for generationComplete
- [b/414602833](https://b.corp.google.com/issues/414602833) -> Fix
firebase ai changelog