Add Firebase AI to recipe version generation (#6951)
Per [b/416734504](https://b.corp.google.com/issues/416734504),

This updates the `GenerateTutorialBundleTask` that we use to generate
the tutorial bundle versions (recipe versions) to support `firebase-ai`.
More specifically, it replaces the existing `firebase-vertexai` entry in
the process.

While this PR removes `firebase-vertexai` from the tutorial bundle, it
does _not_ remove it from the BoM. That will be a breaking change we'll
make down the road.