[action] Only include merge message on sdk changes (#6955)
The "Merge to main" workflow adds a comment to the PR so authors know
that their changes shouldn't break the main branch, as it should always
remain stable and releasable. This only makes sense when the change
affects an SDK, as only those get released. Now we filter the message so
it doesn't spam PRs that only touch infra (plugin & github folders)