Use the same tag for all logs in sessions (#7002)
Use the same tag for all logs in sessions. This will make it easier to
dogfood, and is consistent with Perf and Crashlytics

If any context is lost by losing the tag named after the class, we can
update those log lines as we run into them. A future change we can do is
refactor to use the firebase common logger