Fix cold starts incorrectly using the previous firstSessionId (#7047)
Fix cold starts incorrectly using the previous firstSessionId and
sessionIndex

This came up during the dogfooding session

When a cold start happens, we must always create a new session. Before
this fix, it would preserve the current session's firstSesisonId and
sessionIndex when that session hadn't expired. But a cold start is
regardless of the current session expiring