[FAL] Add support for limited-use tokens (#7285)
Per [b/440338679](https://b.corp.google.com/issues/440338679),

This adds support for the `useLimitedUseAppCheckTokens` parameter to the
`getInstance` constructor and the `.ai` helper method.

To help assist this process, I've also update the
`FirebaseAIMultiResourceComponent` to use a hashed data class containing
all the unique data of each `getInstance`. Previously, it only tracked
instances by the location. Now, it tracks instances by the location,
backend, and the value of `useLimitedUseAppCheckTokens`. In all
practicality, this is unlikely to get much usage, but it helps avoid any
potential edge-case issues where customers expect different instances.

Documentation is present for the added changes, as well as a changelog
entry. These both match what we did on the iCore side, but feel free to
adjust as needed. Furthermore, the `api.txt` file has been updated with
the corresponding changes.