[Infra] Fix broken make change log  (#7325)
The introduction of the markdown formatter accidentally changed the
golden files for the change log generator. Additionally, the new
formatter introduces top-level lists using `-` instead of `*` as we were
using, which caused the generator to fail.

A couple of changes to the matchers were also included as a clean up
step.

Internal b/442932777