[AI] Ignore, and log, unknown parts. (#7333)
The previoius approach, throwing an exception, was based on the
assumption that unknown parts would be important parts of the response.
That would mean that new parts would force devs to update to SDK
versions supporting them before using them. Actually, some times the
parts returned by the model can be safely ignored.

This change modifies the behaviour to Log, as a warning, and ignore,
rather than fail, if an unknown part is found when parsing the response.

Internal b/441783876