[Infra] Annotate subprojects with the type of project
We extended the format of `subprojects.cfg` to support inline
comments. Then, added a comment to indicate what type of project each
entry points to, e.g. SDK or test-only. This is useful for tooling,
like Gemini, as it can now understand better which subprojects to
read/modify if a global, SDK-specific, action is necessary.

The `agents.md` doc has been updated accordingly.