[ALF] Deprecate `UsageMetadata` public constructor
The class was never meant to be instantiated by clients. The public
constructor is now marked as deprecated and will be removed in a
upcoming major version bump.

Additionally, the PR includes an unused parameter in the internal
version to get the code to build. It will be removed in a upcoming
change were an actual new parameter, exclusive to the internal
constructor, is added.