[Infra] Use `-` instead of `*` when moving changelog entries around (#7462)
The merge back action uses the code in \`Changelog.kt\` to ensure  
unreleased entries in the CHANGELOG.md files are moved to the  
corresponding section after being released.  

Since we have started formatting our markdown files, we rely on \`-\` as
a list marker, rather than \`\*\`, which was the old format.  

This change updates \`Changelog.kt\` to use the right character.