SQLiteDatabaseExtsUnitTest.kt, SQLiteArbs.kt, SQLiteArbsUnitTest.kt: Add tests for execSQL with apostrophes in string arguments
The execSQL logging extension function was not being tested with string
arguments containing apostrophes. This could lead to bugs where apostrophes
are not properly escaped, causing SQL syntax errors.

This change introduces property-based tests to cover this case. To do this,
it adds Arbs for generating strings with apostrophes for use in the tests.

* SQLiteDatabaseExtsUnitTest.kt:
    * Add tests for `execSQL` with string bind arguments that contain apostrophes.
* SQLiteArbs.kt:
    * Add Kotest `Arb`s for generating strings that contain apostrophes for property-based testing.
* SQLiteArbsUnitTest.kt:
    * Add unit tests for `SQLiteArbs.kt`.