SQLiteDatabaseExts.kt, SQLiteArbs.kt, SQLiteArbsUnitTest.kt: Improve logging of byte arrays
Improves the logging of byte arrays that are passed as arguments to SQLite queries by logging them as Base64-encoded strings. This is more useful than the previous behavior, which was to log the result of calling `contentToString()`.

Also adds the ability to generate `ByteArray`s to `SQLiteArbs` so that we can test this new logging behavior.

### Changes
* SQLiteDatabaseExts.kt:
    * Log `ByteArray` arguments as Base64-encoded strings.
* SQLiteArbs.kt:
    * Add `byteArrayColumnValue` to generate `ByteArray` column values.
    * Override `equals()`, `hashCode()`, and `toString()` in `ColumnValue`.
* SQLiteArbsUnitTest.kt:
    * Add a test for `byteArrayColumnValue`.