DataConnectCacheDatabaseMigrator.kt: Refactor DataConnectCacheDatabaseMigrator from an object to a class
Refactors the `DataConnectCacheDatabaseMigrator` from a singleton `object` to a `class` with a `companion object`. This change improves the code's structure by allowing the `SQLiteDatabase` and `Logger` to be passed in as constructor parameters and stored as instance variables. This removes the need to pass them as parameters to each method, simplifying the method signatures and making the code more object-oriented.

### Changes
* DataConnectCacheDatabaseMigrator.kt:
    * Converted `DataConnectCacheDatabaseMigrator` from an `object` to a `class`.
    * Added a private constructor that accepts `SQLiteDatabase` and `Logger` instances.
    * Introduced a `companion object` to hold the static `migrate` method and constants.
    * Updated methods to use the instance variables for the database and logger.
    * Changed nested exception classes to be `private`.