fix(ai): Add audio dispatcher (#7483)
Adds an audio dispatcher for dispatching the recording of the microphone
and the playback of audio to separate threads with elevated priorities.
The threads are also marked with `detectNetwork` to catch improper
usage.

This should help avoid weird deadlocks with coroutines and provide a
smoother recording/playback experience in apps with higher thread
traffic.