AudioHandler callback to modify AudioTrack and AudioRecord (#7496)
This PR adds a callback function `audioHandler` which would apply the
necessary configurations or modifications to the `AudioTrack` and the
`AudioRecord` objects used by the `startAudioCoversation` function.

This PR also adds a new configuration class called conversation config
which could be sent to the `startAudioConversation` function which
allows users to specify different aspects of the conversation.