dataconnect: fix `updateJson` task to find Data Connect executables whose file name is updated to also include the cpu architecture.
This naming change started with Data Connect executable version 2.16.0,
which started to be published in both x86_64 and arm64 variants. For
example, the file name for the macos emulator version 2.15.1 was
`dataconnect-emulator-macos-v2.15.1` but in 2.16.0 there were two
executables named `dataconnect-emulator-macos-amd64-v2.16.0` and
`dataconnect-emulator-macos-arm64-v2.16.0` (the only difference being
one is named "amd64" and the other "arm64").